Move tag dtos to TimeScout.Application
